+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## 0.6.1 (unreleased)
 
 * `ComponentOrNode` moved to test module. Renamed to `ReactOrDomNode`.
+* Changed overloaded `classSet` methods into `classSet{,1}{,M}`.
 
 # 0.6.0 ([commit log](https://github.com/japgolly/scalajs-react/compare/v0.5.4...v0.6.0))